feat: expand renderer and community runtime
- Rework the GTA IV native renderer with persistent resources, state/descriptor transport, synchronization, resolve and filtering policy, diagnostics, and CPU/GPU/memory profiling.

- Extend the shader cache by 24 runtime-captured shaders without removing or renaming existing entries.

- Add 64-player guest hooks plus community session, relay, social, stats, progression, chat, voice, and server infrastructure.

- Expand touch, mouse, controller, and rumble input; harden source inspection and installation; and track supporting tests, audits, documentation, captures, and research assets.

+++ glue/rexglue-sdk-main/include/rex/ui/imgui_drawer.h
@@ -110,7 +110,7 @@ class ImGuiDrawer : public WindowInputListener, public UIDrawer {
   // If there's an active pointer, the ImGui mouse is controlled by this touch.
   // If it's TouchEvent::kPointerIDNone, the ImGui mouse is controlled by the
   // mouse.
-  uint32_t touch_pointer_id_ = TouchEvent::kPointerIDNone;
+  uint64_t touch_pointer_id_ = TouchEvent::kPointerIDNone;
   // Whether after the next frame (since the mouse up event needs to be handled
   // with the correct mouse position still), the ImGui mouse position should be
   // reset (for instance, after releasing a touch), so it's not hovering over
